Fix NewBallOnSelect
Previous version would make the NewBall UI element become the parent to the ball on pickup, instead of MainCamera.
This was replaced by calling the player script, which in itself is probably not the cleanest way to do it, but should work.

NOTE: all public variables for player, ball etc. should be replaced by calling a GameManager object, which should be implemented soon.