UI will now spawn when the player first clicks in the game.
The first UI spawn is now based on wheather the UI GameObject is
active/inactive in the initial scene. If the GameObject is set to active
it will spawn with the game to the right of the player. If the
GameObject is set to inactive the UI will be spawned when the player
first uses the click gesture.