Implement simulating AirTap via Mouse
The GazeGestzreRecogniser's Tapped event now calls a method HandleAirTap.
This allows for easy simulation, by triggering it when the Left Mouse Button is clicked.

This and implementing Keyboard movement and Mouse rotation allows testing of some functionalities, without the need to have the HoloLens device nearby, allowing more remote development.
Testing on device must still take place before merging. PlayerKeyboardController and CharacterController should be set to invisible when testing and building to device.